+++ R/new_BayesMixture.R
@@ -1,6 +1,8 @@
 #' Creating a S3 object of class `BayesMixture`
 #' 
 #' Function for creating an object of class `bayes_estimation()` which can subsequently be used as argument in `bayes_mode()`.
+#' This function is useful for users who want to use the mode inference functions of the package with MCMC output generated using 
+#' other software packages.
 #' 
 #' @param mcmc A matrix of MCMC draws
 #' @param data A vector containing the data used for estimating the model and generating the MCMC draws
@@ -30,6 +32,30 @@
 #' @importFrom stringr str_to_lower
 #' @importFrom stringr str_replace
 #' @importFrom stringr str_locate
+#' 
+#' @examples
+#' 
+#' # Example with a Student t ================================================
+#' mu = c(0.5,6)
+#' sigma = c(1,2)
+#' nu = c(5,5)
+#' p = c(0.8,0.2)
+#' params = c(eta = p, mu = mu, sigma = sigma, nu = nu)
+#' pars_names = c("eta", "mu", "sigma", "nu")
+#' dist_type = "continuous"
+#'
+#' data = c(sn::rst(p[1]*1000, mu[1], sigma[1], nu = nu[1]),
+#'          sn::rst(p[2]*1000, mu[2], sigma[2], nu = nu[2]))
+#'
+#' fit = c(eta = p, mu = mu, sigma = sigma, nu = nu)
+#' fit = rbind(fit, fit)
+#' 
+#' pdf_func = function(x, pars) {
+#'   sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
+#' }
+#' 
+#' BM = new_BayesMixture(fit, data, K = 2, burnin = 1,
+#' pars_names = pars_names, pdf_func = pdf_func, dist_type = dist_type)
 #' 
 #' @export
 
@@ -118,7 +144,7 @@ new_BayesMixture <- function(mcmc,
   }
 
   mcmc_all = mcmc_new
-  mcmc = mcmc_all[(burnin+1):nrow(mcmc_all), ]
+  mcmc = mcmc_all[(burnin+1):nrow(mcmc_all), ,drop = FALSE]
   
   
   if (dist %in% c("normal", "skew_normal",
@@ -127,6 +153,8 @@ new_BayesMixture <- function(mcmc,
   } else {
     BayesMix$dist = "NA"
     BayesMix$pdf_func = pdf_func
+    assert_that(!is.null(pdf_func),
+                msg = "pdf_func is missing")
   }
   
   BayesMix$pars_names = pars_names