+++ tests/testthat/test-new_BayesMixture.R
@@ -22,7 +22,6 @@ test_that("New_BayesMixture works returns expected results for the student distr
   p = c(0.8,0.2)
   params = c(eta = p, mu = mu, sigma = sigma, nu = nu)
   pars_names = c("eta", "mu", "sigma", "nu")
-  dist = "student"
   dist_type = "continuous"
   
   data = c(sn::rst(p[1]*1000, mu[1], sigma[1], nu = nu[1]),
@@ -31,6 +30,35 @@ test_that("New_BayesMixture works returns expected results for the student distr
   fit = c(eta = p, mu = mu, sigma = sigma, nu = nu)
   fit = rbind(fit, fit)
   
-  BM = new_BayesMixture(fit, data, K = 2, burnin = 1, dist = dist, pars_names = pars_names, pdf_func = NULL, dist_type = dist_type)
+  pdf_func = function(x, pars) {
+    sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
+  }
+  
+  BM = new_BayesMixture(fit, data, K = 2, burnin = 1, pars_names = pars_names, pdf_func = pdf_func, dist_type = dist_type)
   expect_s3_class(BM, "BayesMixture")
+})
+
+test_that("New_BayesMixture gives expected error", {
+  set.seed(1)
+  mu = c(0.5,6)
+  sigma = c(1,2)
+  nu = c(5,5)
+  p = c(0.8,0.2)
+  params = c(eta = p, mu = mu, sigma = sigma, nu = nu)
+  pars_names = c("eta", "mu", "omega", "nu")
+  dist_type = "continuous"
+  
+  data = c(sn::rst(p[1]*1000, mu[1], sigma[1], nu = nu[1]),
+           sn::rst(p[2]*1000, mu[2], sigma[2], nu = nu[2]))
+  
+  fit = c(eta = p, mu = mu, sigma = sigma, nu = nu)
+  fit = rbind(fit, fit)
+  
+  pdf_func = function(x, pars) {
+    sn::dst(x, pars["mu"], pars["sigma"], pars["xi"], pars["nu"])
+  }
+  
+  expect_error(new_BayesMixture(fit, data, K = 2, burnin = 1, pars_names = pars_names,
+                                pdf_func = pdf_func, dist_type = dist_type),
+               "the name of the parameters provided by pars_names and those of the mcmc object do not match")
 })